Move CaptureSelfieActivity to new module
