+++ selfie-camera/build.gradle.kts
@@ -44,7 +44,10 @@ android {
 }
 
 dependencies {
+    implementation(project(":shared"))
     implementation(project(":androidshared"))
+    implementation(project(":strings"))
+    implementation(project(":permissions"))
 
     implementation(Dependencies.timber)
 }
