+++ selfie-camera/src/main/res/layout/activity_capture_selfie.xml
